# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Опора АПК",
    "description": "Навигатор поддержки. Подбор предварительный. Локальный набор — синтетический.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/auth/demo": {
      "post": {
        "summary": "Demo Login",
        "operationId": "demo_login_api_auth_demo_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DemoInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionView"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          },
          "403": {
            "description": "Недостаточно прав или запрос не прошёл проверку защиты"
          }
        },
        "parameters": [
          {
            "name": "X-App-Request",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "const": "1"
            },
            "description": "Защита запросов приложения. Origin должен соответствовать PUBLIC_ORIGIN."
          }
        ]
      }
    },
    "/api/auth/max": {
      "post": {
        "summary": "Max Login",
        "operationId": "max_login_api_auth_max_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MaxInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionView"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          },
          "403": {
            "description": "Недостаточно прав или запрос не прошёл проверку защиты"
          }
        },
        "parameters": [
          {
            "name": "X-App-Request",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "const": "1"
            },
            "description": "Защита запросов приложения. Origin должен соответствовать PUBLIC_ORIGIN."
          }
        ]
      }
    },
    "/api/auth/me": {
      "get": {
        "summary": "Me",
        "operationId": "me_api_auth_me_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionView"
                }
              }
            }
          },
          "401": {
            "description": "Нет действующей сессии"
          },
          "403": {
            "description": "Недостаточно прав или запрос не прошёл проверку защиты"
          }
        },
        "security": [
          {
            "sessionCookie": []
          }
        ]
      }
    },
    "/api/auth/logout": {
      "post": {
        "summary": "Logout",
        "operationId": "logout_api_auth_logout_post",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "401": {
            "description": "Нет действующей сессии"
          },
          "403": {
            "description": "Недостаточно прав или запрос не прошёл проверку защиты"
          }
        },
        "security": [
          {
            "sessionCookie": [],
            "csrfHeader": []
          }
        ],
        "parameters": [
          {
            "name": "X-App-Request",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "const": "1"
            },
            "description": "Защита запросов приложения. Origin должен соответствовать PUBLIC_ORIGIN."
          }
        ]
      }
    },
    "/api/profile": {
      "get": {
        "summary": "Profile",
        "operationId": "profile_api_profile_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BusinessProfile-Output"
                }
              }
            }
          },
          "401": {
            "description": "Нет действующей сессии"
          },
          "403": {
            "description": "Недостаточно прав или запрос не прошёл проверку защиты"
          }
        },
        "security": [
          {
            "sessionCookie": []
          }
        ]
      },
      "put": {
        "summary": "Save Profile",
        "operationId": "save_profile_api_profile_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BusinessProfile-Input"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BusinessProfile-Output"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          },
          "401": {
            "description": "Нет действующей сессии"
          },
          "403": {
            "description": "Недостаточно прав или запрос не прошёл проверку защиты"
          }
        },
        "security": [
          {
            "sessionCookie": [],
            "csrfHeader": []
          }
        ],
        "parameters": [
          {
            "name": "X-App-Request",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "const": "1"
            },
            "description": "Защита запросов приложения. Origin должен соответствовать PUBLIC_ORIGIN."
          }
        ]
      }
    },
    "/api/profile/questions": {
      "get": {
        "summary": "Profile Questions",
        "operationId": "profile_questions_api_profile_questions_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ProfileQuestion"
                  },
                  "type": "array",
                  "title": "Response Profile Questions Api Profile Questions Get"
                }
              }
            }
          },
          "401": {
            "description": "Нет действующей сессии"
          },
          "403": {
            "description": "Недостаточно прав или запрос не прошёл проверку защиты"
          }
        },
        "security": [
          {
            "sessionCookie": []
          }
        ]
      }
    },
...
    "/api/measures/{measure_id}": {
      "get": {
        "summary": "Measure Detail",
        "operationId": "measure_detail_api_measures__measure_id__get",
        "parameters": [
          {
            "name": "measure_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Measure Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeasureView"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/matches": {
      "post": {
        "summary": "Matches",
        "operationId": "matches_api_matches_post",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MatchesView"
                }
              }
            }
          },
          "401": {
            "description": "Нет действующей сессии"
          },
          "403": {
            "description": "Недостаточно прав или запрос не прошёл проверку защиты"
          }
        },
        "security": [
          {
            "sessionCookie": [],
            "csrfHeader": []
          }
        ],
        "parameters": [
          {
            "name": "X-App-Request",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "const": "1"
            },
            "description": "Защита запросов приложения. Origin должен соответствовать PUBLIC_ORIGIN."
          }
        ]
      }
    },
    "/api/preparation-plans": {
      "get": {
        "summary": "Plans",
        "operationId": "plans_api_preparation_plans_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/PlanView"
                  },
                  "type": "array",
                  "title": "Response Plans Api Preparation Plans Get"
                }
              }
            }
          },
          "401": {
            "description": "Нет действующей сессии"
          },
          "403": {
            "description": "Недостаточно прав или запрос не прошёл проверку защиты"
          }
        },
        "security": [
          {
            "sessionCookie": []
          }
        ]
      },
      "post": {
        "summary": "Create Plan",
        "operationId": "create_plan_api_preparation_plans_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PlanInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlanView"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          },
          "401": {
            "description": "Нет действующей сессии"
          },
          "403": {
            "description": "Недостаточно прав или запрос не прошёл проверку защиты"
          }
        },
        "security": [
          {
            "sessionCookie": [],
            "csrfHeader": []
          }
        ],
        "parameters": [
          {
            "name": "X-App-Request",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "const": "1"
            },
            "description": "Защита запросов приложения. Origin должен соответствовать PUBLIC_ORIGIN."
          }
        ]
      }
    },
    "/api/preparation-plans/{plan_id}/items/{item_id}": {
      "patch": {
        "summary": "Check Item",
        "operationId": "check_item_api_preparation_plans__plan_id__items__item_id__patch",
        "parameters": [
          {
            "name": "plan_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Plan Id"
            }
          },
          {
            "name": "item_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Item Id"
            }
          },
          {
            "name": "X-App-Request",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "const": "1"
            },
            "description": "Защита запросов приложения. Origin должен соответствовать PUBLIC_ORIGIN."
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CheckInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlanView"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          },
          "401": {
            "description": "Нет действующей сессии"
          },
          "403": {
            "description": "Недостаточно прав или запрос не прошёл проверку защиты"
          }
        },
        "security": [
          {
            "sessionCookie": [],
            "csrfHeader": []
          }
        ]
      }
    },
    "/api/admin/versions": {
      "get": {
        "summary": "Admin Versions",
        "operationId": "admin_versions_api_admin_versions_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/MeasureView"
                  },
                  "type": "array",
                  "title": "Response Admin Versions Api Admin Versions Get"
                }
              }
            }
          },
          "401": {
            "description": "Нет действующей сессии"
          },
          "403": {
            "description": "Недостаточно прав или запрос не прошёл проверку защиты"
          }
        },
        "security": [
          {
            "sessionCookie": []
          }
        ]
      }
    },
    "/api/admin/preview": {
      "post": {
        "summary": "Preview Rules",
        "operationId": "preview_rules_api_admin_preview_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PreviewInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PreviewView"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          },
          "401": {
            "description": "Нет действующей сессии"
          },
          "403": {
            "description": "Недостаточно прав или запрос не прошёл проверку защиты"
          }
        },
        "security": [
          {
            "sessionCookie": [],
            "csrfHeader": []
          }
        ],
        "parameters": [
          {
            "name": "X-App-Request",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "const": "1"
            },
            "description": "Защита запросов приложения. Origin должен соответствовать PUBLIC_ORIGIN."
          }
        ]
      }
    },
...
    "/api/max/webhook": {
      "post": {
        "summary": "Webhook",
        "operationId": "webhook_api_max_webhook_post",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "403": {
            "description": "Недостаточно прав или запрос не прошёл проверку защиты"
          }
        },
        "security": [
          {
            "maxWebhookSecret": []
          }
        ]
      }
    }
  },
  "components": {
    "schemas": {
...
      "DemoInput": {
        "properties": {
          "persona": {
            "type": "string",
            "enum": [
              "farmer",
              "editor",
              "second"
            ],
            "title": "Persona",
            "default": "farmer"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "DemoInput"
      },
...
      "SessionView": {
        "properties": {
          "user": {
            "$ref": "#/components/schemas/UserView"
          },
          "csrf": {
            "type": "string",
            "title": "Csrf"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "user",
          "csrf"
        ],
        "title": "SessionView"
      },
...
    },
    "securitySchemes": {
      "sessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "opora_session",
        "description": "HttpOnly cookie после /api/auth/max; /api/auth/demo только локально"
      },
      "csrfHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "X-CSRF-Token",
        "description": "Значение csrf из входа или /api/auth/me"
      },
      "maxWebhookSecret": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Max-Bot-Api-Secret"
      }
    }
  },
  "servers": [
    {
      "url": "/",
      "description": "Текущий сервер"
    }
  ]
}